requirements.txt
flask
openai-whisper
sentence-transformers
faiss-cpu
torch
torchaudio